Removed ObsoleteAttribute from the old methods' signatures and moved the AddPicture methods from the OpenXmlExporter to the OpenXmlTemplater
