+++ src/MiniExcel.Core/Api/OpenXmlExporter.cs
@@ -1,5 +1,3 @@
-using MiniExcelLib.Core.OpenXml.Picture;
-
 // ReSharper disable once CheckNamespace
 namespace MiniExcelLib.Core;
 
@@ -8,19 +6,6 @@ public sealed partial class OpenXmlExporter
     internal OpenXmlExporter() { }
     
     
-    [CreateSyncVersion]
-    public async Task AddPictureAsync(string path, CancellationToken cancellationToken = default, params MiniExcelPicture[] images)
-    {
-        using var stream = File.Open(path, FileMode.OpenOrCreate);
-        await MiniExcelPictureImplement.AddPictureAsync(stream, cancellationToken, images).ConfigureAwait(false);
-    }
-
-    [CreateSyncVersion]
-    public async Task AddPictureAsync(Stream excelStream, CancellationToken cancellationToken = default, params MiniExcelPicture[] images)
-    {
-        await MiniExcelPictureImplement.AddPictureAsync(excelStream, cancellationToken, images).ConfigureAwait(false);
-    }
-
     [CreateSyncVersion]
     public async Task<int> InsertSheetAsync(string path, object value, string? sheetName = "Sheet1", bool printHeader = true, bool overwriteSheet = false, OpenXmlConfiguration? configuration = null, CancellationToken cancellationToken = default)
     {
